# swaggerdocumentation.yaml
openapi: 3.0.0
info:
  title: BlogApp
  version: "1.0.0"
  description: |
    API documentation for the BlogApp application.
  contact:
    name: Imanariyo Baptiste
    email: imanariyobaptiste@gmail.com
    telphone: "+250787795163"

    url: "https://imanariyo-portfolio-web.vercel.app/"
servers:
  - url: http://localhost:3330
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /seed:
    post:
      tags:
        - Seed
      summary: Seed the database
      description: Seeds the database with sample data.
      responses:
        '200':
          description: Database seeded successfully.
        '500':
          description: Error seeding database.
  /sync-db:
    post:
      tags:
        - Sync
      summary: Sync the database
      description: 
      # Synchronizes the database schema with the models. 
        This operation will:
        - Drop and recreate all tables if using `{ force: true }`. 
          # This results in the loss of existing data.
        - Apply any new changes to the schema as defined in your models.
        - Ensure the database structure matches your Sequelize models.
        - It's important to use this operation cautiously in production environments to avoid unintentional data loss.
      responses:
        '200':
          description: 
            Database synced successfully. The database schema has been updated to match the models.
        '500':
          description: 
            Error syncing database. An issue occurred while trying to synchronize the database schema, such as connectivity problems or schema definition errors.
...
  /posts:
    get:
      tags:
        - Posts
      summary: Get All Posts
      description: Retrieves a list of all posts.
      responses:
        '200':
          description: Successful retrieval
        '404':
          description: Posts not found

    post:
      tags:
        - Posts
      summary: Create Post
      description: Creates a new post.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: My First Post
                content:
                  type: string
                  example: This is the content of the first post.
                authorId:
                  type: integer
                  example: 1
                image:
                  type: string
                  format: binary
              required:
                - title
                - content
      responses:
        '201':
          description: Post created successfully
        '400':
          description: Invalid request data
...